Gallery2: if phone is locked, pause the video, if not, start it.
If there was a video playing before locking the phone,
After pressing the emergency call option in lock screen,
and then presses back key, the MovieActivity will resume,
so the video will play background.

Set a listener, if Keyguard is locked,pause the video,
if not, start it.

CRs-Fixed: 427283

Change-Id: Ia825afee691ee53f53eebf1f96efe09ef88ca743

Conflicts:
	src/com/android/gallery3d/app/MoviePlayer.java